# docs/swagger.yaml
basePath: /api/v1
definitions:
  api.auth:
    properties:
      password:
        type: string
      username:
        type: string
    type: object
  v1.createArticle:
    properties:
      body:
        type: string
      channelId:
        type: integer
      title:
        type: string
    type: object
  v1.tagCreate:
    properties:
      name:
        example: account name
        type: string
      state:
        example: 0
        type: integer
    type: object
info:
  contact: {}
  description: This is a sample server Petstore server.
  license: {}
  title: Swagger API for gsgo
  version: "2.0"
paths:
  /articles:
    post:
      parameters:
      - description: 新建
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/v1.createArticle'
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: AddArticle
      tags:
      - 文章
  /articles/{id}:
    get:
      consumes:
      - application/json
      description: get article by ID
      operationId: tagId
      parameters:
      - description: aticle ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: get a article by id
      tags:
      - 文章
  /auth:
    post:
      parameters:
      - description: 登录
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/api.auth'
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Get Auth
      tags:
      - auth
  /channels:
    get:
      description: get all channel
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: get channels
      tags:
      - 频道
  /tags:
    post:
      parameters:
      - description: 新建
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/v1.tagCreate'
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: AddTag
      tags:
      - 标签
  /tags/:
    get:
      consumes:
      - application/json
      description: get tags list
      parameters:
      - description: name
        in: query
        name: name
        type: string
      - description: page
        in: query
        name: page
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          headers:
            Token:
              description: qwerty
              type: string
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: get tags list
      tags:
      - 标签
  /tags/{id}:
    delete:
      consumes:
      - application/json
      description: delete tag by ID
      operationId: tagId
      parameters:
      - description: tag ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: delete a tag
      tags:
      - 标签
    get:
      consumes:
      - application/json
      description: put tag by ID
      operationId: tagId
      parameters:
      - description: tag ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: put a tag
      tags:
      - 标签
    put:
      consumes:
      - application/json
      description: get tag by ID
      operationId: tagId
      parameters:
      - description: tag ID
        in: path
        name: id
        required: true
        type: integer
      - description: 修改
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/v1.tagCreate'
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: get a tag
      tags:
      - 标签
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"